requirements.txt
flask
flask_sqlalchemy
flask_wtf
wtforms
python-dotenv
mysql-connector-python
prometheus_flask_exporter
